+++ manifests/init_scripts/lib_functions.sh
@@ -1,0 +1,17 @@
+#!/bin/bash
+#
+# lib_functions.sh
+#
+
+#
+# Guard
+# =============
+#
+# if [ $0 != "-bash" ]; then
+#     echo "Do not run this script stand alone!"
+#     echo "<= Exiting."
+#     exit;
+# fi;
+
+USER=vagrant
+CHEF_IP="10.10.10.10"