+++ changelog.md
@@ -3,6 +3,9 @@
 * v0.0.8
   - Add `support` dir, to be used for work arounds and arch specific things on dev host
   - Add `vboxload.sh` to reload network drivers, fixes #13 (current Mavericks + vbox bug)
+  - Add `location` node to `downloads`, which is a key defined in `locations`
+  - Add `track` node to `downloads`, which will track the external resource
+  - Add functionality to track downloads
 
 * v0.0.7
   - Add open source chef server to close #10 (using Ubuntu 12.04)